Delete dumb usecase SetMessageForReply
